+++ src/main/java/com/mia/bookstore/repository/PendingUserRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.Optional;
 
-public interface PendingUserRepository extends JpaRepository<PendingUser, Long> {
+public interface PendingUserRepository extends JpaRepository<PendingUser, Integer> {
 
     Optional<PendingUser> findByActivationCode(String activationCode);
